requirements.txt
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.114.0
paramiko==3.4.0
python-dotenv==1.0.0
